requirements.txt
xstreamlit==0.56.0
pandas==0.25.1
numpy==1.17.2
matplotlib==3.1.1
statsmodels==0.10.1
patsy==0.5.1
xlrd>=1.0.0
scipy==1.4.1